+++ src/graphics/rc_sprite.rs
@@ -15,14 +15,14 @@ const PANIC_ERROR_MSG: &str = "Sprite does not hold a texture! Return value cann
 
 /// Drawable representation of a texture
 ///
-/// Sprite is a drawable type that allows to easily
+/// `RcSprite` is a drawable type that allows to easily
 /// display a [`RcTexture`] (or a part of it) on a render target.
 ///
 /// __Note:__
 /// This is an improvement upon the Sprite module which dissallows seperation from the texture
-/// lifetime. The `RcSprite` allows from complete seperation from the `RcTexture` while still
-/// referencing it. Underneath, it uses reference counting to ensure that the `RcTexture` is alive and
-/// well, and will throw errors messages if you try to perform function on the sprite while the
+/// lifetime. The `RcSprite` allows for complete seperation from the `RcTexture` while still
+/// referencing it. Underneath, it uses reference counting to ensure that the `RcTexture` is alive,
+/// and will throw errors messages if you try to perform function on the sprite while the
 /// `RcTexture` is no longer alive.
 #[derive(Debug)]
 pub struct RcSprite {
@@ -72,12 +72,13 @@ impl RcSprite {
     /// doesn't store its own copy of the texture, but rather keeps
     /// a pointer to the one that you passed to this function.
     /// If the source texture is destroyed and the sprite tries to
-    /// use it, the behaviour is undefined.
+    /// use it, the behavior is defined within `RcSprite`. `RcSprite` will either do nothing,
+    /// panic, or return a default implementation for specific functions.
     /// If `reset_rect` is true, the [`texture_rect`] property of
     /// the sprite is automatically adjusted to the size of the new
     /// texture. If it is false, the texture rect is left unchanged.
     ///
-    /// [`texture_rect`]: Sprite::texture_rect
+    /// [`texture_rect`]: RcSprite::texture_rect
     ///
     /// # Arguments
     /// * `texture` - New texture
@@ -100,6 +101,7 @@ impl RcSprite {
     /// texture. It can be used to colorize the sprite, or change
     /// its global opacity.
     /// By default, the sprite's color is opaque white.
+    /// Function fails and prints an error message if `RcSprite`'s texture is dead.
     ///
     /// # Arguments
     /// * color - New color of the sprite
@@ -130,6 +132,7 @@ impl RcSprite {
     /// Get the global color of a sprite
     ///
     /// Return the global color of the sprite
+    /// Function fails, returns default, and prints an error message if `RcSprite`'s texture is dead.
     #[must_use]
     pub fn color(&self) -> Color {
         if !self.texture_exists() {
@@ -147,7 +150,8 @@ impl RcSprite {
     /// In other words, this function returns the bounds of the
     /// entity in the entity's coordinate system.
     ///
-    /// Return the local bounding rectangle of the entity
+    /// Return the local bounding rectangle of the entity.
+    /// Function fails, returns default, and prints an error message if `RcSprite`'s texture is dead.
     #[must_use]
     pub fn local_bounds(&self) -> FloatRect {
         if !self.texture_exists() {
@@ -164,8 +168,10 @@ impl RcSprite {
     /// rotation, scale, ...) that are applied to the entity.
     /// In other words, this function returns the bounds of the
     /// sprite in the global 2D world's coordinate system.
+    /// Function fails, returns default, and prints an error message if `RcSprite`'s texture is dead.
     ///
     /// Return the global bounding rectangle of the entity
+    /// Function fails, returns default, and prints an error message if `RcSprite`'s texture is dead.
     #[must_use]
     pub fn global_bounds(&self) -> FloatRect {
         if !self.texture_exists() {
@@ -176,8 +182,10 @@ impl RcSprite {
     }
 
     /// Get the sub-rectangle of the texture displayed by a sprite
+    /// Function fails, returns default, and prints an error message if `RcSprite`'s texture is dead.
     ///
     /// Return the texture rectangle of the sprite
+    /// Function fails, returns default, and prints an error message if `RcSprite`'s texture is dead.
     #[must_use]
     pub fn texture_rect(&self) -> IntRect {
         if !self.texture_exists() {
@@ -192,6 +200,7 @@ impl RcSprite {
     /// The texture rect is useful when you don't want to display
     /// the whole texture, but rather a part of it.
     /// By default, the texture rect covers the entire texture.
+    /// Function fails, and prints an error message if `RcSprite`'s texture is dead.
     ///
     /// # Arguments
     /// * rectangle - Rectangle defining the region of the texture to display
@@ -216,9 +225,9 @@ impl Default for RcSprite {
 
 impl Clone for RcSprite {
     /// Return a new Sprite or panic! if there is not enough memory
-    fn clone(&self) -> RcSprite {
+    fn clone(&self) -> Self {
         let sp = unsafe { ffi::sfSprite_copy(self.sprite.as_ptr()) };
-        RcSprite {
+        Self {
             sprite: NonNull::new(sp).expect("Failed to copy Sprite"),
             texture: self.texture.clone(),
         }
@@ -238,6 +247,9 @@ impl Drawable for RcSprite {
 }
 
 impl Transformable for RcSprite {
+    /// Reference [`Transformable::set_position`] for additional information
+    ///
+    /// Function fails, and prints an error message if `RcSprite`'s texture is dead.
     fn set_position<P: Into<Vector2f>>(&mut self, position: P) {
         if !self.texture_exists() {
             eprintln!("{ERROR_MSG}");
@@ -245,6 +257,9 @@ impl Transformable for RcSprite {
         }
         unsafe { ffi::sfSprite_setPosition(self.sprite.as_ptr(), position.into()) }
     }
+    /// Reference [`Transformable::set_rotation`] for additional information
+    ///
+    /// Function fails, and prints an error message if `RcSprite`'s texture is dead.
     fn set_rotation(&mut self, angle: f32) {
         if !self.texture_exists() {
             eprintln!("{ERROR_MSG}");
@@ -252,6 +267,9 @@ impl Transformable for RcSprite {
         }
         unsafe { ffi::sfSprite_setRotation(self.sprite.as_ptr(), angle) }
     }
+    /// Reference [`Transformable::set_scale`] for additional information
+    ///
+    /// Function fails, and prints an error message if `RcSprite`'s texture is dead.
     fn set_scale<S: Into<Vector2f>>(&mut self, scale: S) {
         if !self.texture_exists() {
             eprintln!("{ERROR_MSG}");
@@ -259,6 +277,9 @@ impl Transformable for RcSprite {
         }
         unsafe { ffi::sfSprite_setScale(self.sprite.as_ptr(), scale.into()) }
     }
+    /// Reference [`Transformable::set_origin`] for additional information
+    ///
+    /// Function fails, and prints an error message if `RcSprite`'s texture is dead.
     fn set_origin<O: Into<Vector2f>>(&mut self, origin: O) {
         if !self.texture_exists() {
             eprintln!("{ERROR_MSG}");
@@ -266,6 +287,9 @@ impl Transformable for RcSprite {
         }
         unsafe { ffi::sfSprite_setOrigin(self.sprite.as_ptr(), origin.into()) }
     }
+    /// Reference [`Transformable::position`] for additional information
+    ///
+    /// Function fails, returns default, and prints an error message if `RcSprite`'s texture is dead.
     fn position(&self) -> Vector2f {
         if !self.texture_exists() {
             eprintln!("{RETURN_ERROR_MSG}");
@@ -273,6 +297,9 @@ impl Transformable for RcSprite {
         }
         unsafe { ffi::sfSprite_getPosition(self.sprite.as_ptr()) }
     }
+    /// Reference [`Transformable::rotation`] for additional information
+    ///
+    /// Function fails, returns default, and prints an error message if `RcSprite`'s texture is dead.
     fn rotation(&self) -> f32 {
         if !self.texture_exists() {
             eprintln!("{RETURN_ERROR_MSG}");
@@ -280,6 +307,9 @@ impl Transformable for RcSprite {
         }
         unsafe { ffi::sfSprite_getRotation(self.sprite.as_ptr()) }
     }
+    /// Reference [`Transformable::get_scale`] for additional information
+    ///
+    /// Function fails, returns default, and prints an error message if `RcSprite`'s texture is dead.
     fn get_scale(&self) -> Vector2f {
         if !self.texture_exists() {
             eprintln!("{RETURN_ERROR_MSG}");
@@ -287,6 +317,9 @@ impl Transformable for RcSprite {
         }
         unsafe { ffi::sfSprite_getScale(self.sprite.as_ptr()) }
     }
+    /// Reference [`Transformable::origin`] for additional information
+    ///
+    /// Function fails, returns default, and prints an error message if `RcSprite`'s texture is dead.
     fn origin(&self) -> Vector2f {
         if !self.texture_exists() {
             eprintln!("{RETURN_ERROR_MSG}");
@@ -294,6 +327,9 @@ impl Transformable for RcSprite {
         }
         unsafe { ffi::sfSprite_getOrigin(self.sprite.as_ptr()) }
     }
+    /// Reference [`Transformable::move_`] for additional information
+    ///
+    /// Function fails, and prints an error message if `RcSprite`'s texture is dead.
     fn move_<O: Into<Vector2f>>(&mut self, offset: O) {
         if !self.texture_exists() {
             eprintln!("{ERROR_MSG}");
@@ -301,6 +337,9 @@ impl Transformable for RcSprite {
         }
         unsafe { ffi::sfSprite_move(self.sprite.as_ptr(), offset.into()) }
     }
+    /// Reference [`Transformable::rotate`] for additional information
+    ///
+    /// Function fails, and prints an error message if `RcSprite`'s texture is dead.
     fn rotate(&mut self, angle: f32) {
         if !self.texture_exists() {
             eprintln!("{ERROR_MSG}");
@@ -308,6 +347,9 @@ impl Transformable for RcSprite {
         }
         unsafe { ffi::sfSprite_rotate(self.sprite.as_ptr(), angle) }
     }
+    /// Reference [`Transformable::scale`] for additional information
+    ///
+    /// Function fails, and prints an error message if `RcSprite`'s texture is dead.
     fn scale<F: Into<Vector2f>>(&mut self, factors: F) {
         if !self.texture_exists() {
             eprintln!("{ERROR_MSG}");
@@ -315,12 +357,18 @@ impl Transformable for RcSprite {
         }
         unsafe { ffi::sfSprite_scale(self.sprite.as_ptr(), factors.into()) }
     }
+    /// Reference [`Transformable::transform`] for additional information
+    ///
+    /// Panics if texture doesn't exist
     fn transform(&self) -> &Transform {
         if !self.texture_exists() {
             panic!("{}", PANIC_ERROR_MSG);
         }
         unsafe { &*ffi::sfSprite_getTransform(self.sprite.as_ptr()) }
     }
+    /// Reference [`Transformable::inverse_transform`] for additional information
+    ///
+    /// Panics if texture doesn't exist
     fn inverse_transform(&self) -> &Transform {
         if !self.texture_exists() {
             panic!("{}", PANIC_ERROR_MSG);