+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
 ## Unreleased
+
 ### Added
+- Added `RcFont` and `RcText` for reference counted text (see `examples/rc-resources.rs`)
 - pub const DEFAULT for `Vertex`
 
 ### Changed